Separação de código para melhor manutenção
