+++ GerenciamentoVultorian.Api/Config/AppConfiguration.cs
@@ -1,0 +1,19 @@
+﻿namespace GerenciamentoVultorian.Api.Config;
+
+public static class AppConfiguration
+{
+    public static WebApplication ConfigureApp(this WebApplication app)
+    {
+        if (app.Environment.IsDevelopment())
+        {
+            app.UseSwagger();
+            app.UseSwaggerUI();
+        }
+
+        app.UseHttpsRedirection();
+        app.UseAuthorization();
+        app.MapControllers();
+
+        return app;
+    }
+}
