+++ GerenciamentoVultorian.Api/Config/DependencyInjectionSetup.cs
@@ -1,0 +1,13 @@
+﻿namespace GerenciamentoVultorian.Api.Config;
+
+public static class DependencyInjectionSetup
+{
+    public static WebApplicationBuilder ConfigureDependencyInjection(this WebApplicationBuilder builder)
+    {
+        builder.Services.AddControllers();
+        builder.Services.AddEndpointsApiExplorer();
+        builder.Services.AddSwaggerGen();
+
+        return builder;
+    }
+}
